Move the cdsMaturity function to CDS instrument file.
