+++ ql/time/schedule.hpp
@@ -151,20 +151,10 @@ namespace QuantLib {
         Date firstDate_, nextToLastDate_;
     };
 
-    /*! Return the CDS maturity date given the CDS trade date, \p tradeDate, the CDS \p tenor and a CDS \p rule.
-
-        A \c Null<Date>() is returned when a \p rule of \c CDS2015 and a \p tenor length of zero fail to yield a valid 
-        CDS maturity date.
-
-        \warning An exception will be thrown if the \p rule is not \c CDS2015, \c CDS or \c OldCDS.
-
-        \warning An exception will be thrown if the \p rule is \c OldCDS and a \p tenor of 0 months is provided. This 
-                 restriction can be removed if 0M tenor was available before the CDS Big Bang 2009.
-
-        \warning An exception will be thrown if the \p tenor is not a multiple of 3 months. For the avoidance of 
-                 doubt, a \p tenor of 0 months is supported.
+    /*! Helper function for returning the date on or before date \p d that is the 20th of the month and obeserves the 
+        given date generation \p rule if it is relevant.
     */
-    Date cdsMaturity(const Date& tradeDate, const Period& tenor, DateGeneration::Rule rule);
+    Date previousTwentieth(const Date& d, DateGeneration::Rule rule);
 
     // inline definitions
